COMP: Remove explicit instantiations of VectorImage from ImageFileReader and ImageFileWriter. The only reason they were needed was to access a method available in the VectorImage and not in Image. We now use PixelAccessorFunctors to SetVectorLength while reading an image and GetVectorLength while writing an image. For consistency the DefaultPixelAccessorFunctor has a dummy method Set/GetVectorLength(). This should hopefully solve the build problems on SGI, ICC etc

+++ Code/Common/itkDefaultVectorPixelAccessor.h
@@ -49,7 +49,7 @@ class ITK_EXPORT DefaultVectorPixelAccessor
 {
 public:
 
-  typedef unsigned long VectorLengthType;
+  typedef unsigned int VectorLengthType;
   
   /** External typedef. It defines the external aspect
    * that this class will exhibit. Here it is an Array. The container does not
